[User] Implement the ability to display user's reviews

+++ app/src/main/java/gr/fellow/fellow_traveller/domain/mappers/ReviewMapper.kt
@@ -4,5 +4,5 @@ import gr.fellow.fellow_traveller.domain.review.Review
 import gr.fellow.fellow_traveller.framework.network.fellow.review.ReviewResponse
 
 fun ReviewResponse.mapReview() = Review(
-    user.mapToUserBase(), rate, timestamp
+    user.mapToUserBase(), rate.toInt(), timestamp
 )
