+++ app/src/main/java/gr/fellow/fellow_traveller/domain/review/Review.kt
@@ -4,6 +4,6 @@ import gr.fellow.fellow_traveller.domain.user.UserBase
 
 data class Review(
     val user: UserBase,
-    val rate: Float,
+    val rate: Int,
     val timestamp: Long
 )
